+++ src/fwtv/widgets/login_widget.py
@@ -1,3 +1,4 @@
+from PySide6.QtCore import QDate
 from PySide6.QtCore import Qt
 from PySide6.QtGui import QKeyEvent
 from PySide6.QtWidgets import QHBoxLayout
@@ -7,11 +8,24 @@ from PySide6.QtWidgets import QPushButton
 from PySide6.QtWidgets import QSizePolicy
 from PySide6.QtWidgets import QWidget
 
+from fwtv.widgets import settings_widget
+
 
 class LoginWidget(QWidget):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.qh = QHBoxLayout(self)
+
+        self.fetch_data_label = QLabel("Fetch data", self)
+        self.fetch_data_label.setSizePolicy(QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Fixed)
+        self.qh.addWidget(self.fetch_data_label)
+
+        # defaults to fetching data from the last 2 months in order to not fetch entire data history
+        self.start_picker = settings_widget.DateSettingWidget("From", QDate.currentDate().addMonths(-2), self)
+        self.qh.addWidget(self.start_picker)
+
+        self.end_picker = settings_widget.DateSettingWidget("Until", QDate.currentDate(), self)
+        self.qh.addWidget(self.end_picker)
 
         self.label = QLabel("Enter api key", self)
         self.label.setSizePolicy(QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Fixed)
@@ -24,6 +38,9 @@ class LoginWidget(QWidget):
         self.button = QPushButton("Fetch data", self)
         self.button.setSizePolicy(QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Fixed)
         self.qh.addWidget(self.button)
+
+        self.timeout = settings_widget.IntegerPickerWidget("Timeout", 5)
+        self.qh.addWidget(self.timeout)
 
         self.setLayout(self.qh)